ASoC: wcd-dsp-mgr: add suspend and resume functionality
Currently, the suspend and resume routines for the wcd dsp manager
driver are not implemented, which causes race conditions when the
components suspend or resume out of order. Implement the suspend/
resume functionality and route the events to the components in
the expected sequence.

CRs-Fixed: 1111863
Change-Id: Ie84cdef5ed7ffd61610cd83e5f627427f85bc125
Signed-off-by: Bhalchandra Gajare <gajare@codeaurora.org>
Signed-off-by: Eliminater74 <eliminater74@gmail.com>
(cherry picked from commit 1d03567c374655a95d54114eb56aab98c75e0d96)