Avoid a second dispatch in case of budget exhaustion
IMPROVEMENT This commit anticipates the complete check of budget exhaustion,
for the in-service bfq_queue, to when the next bfq_queue to serve is selected
(during a dispatch operation). This enables a new bfq_queue to be immediately
selected for service in case the in-service bfq_queue has actually exhausted
its budget. As a consequence, a second dispatch invocation is not needed
any more, to have a new request dispatched.
To implement this improvement, this commit implements a further improvement
too: the field next_rq of a bfq_queue now always contains the actual next
request to dispatch (or NULL if the bfq_queue is empty).

Signed-off-by: Paolo Valente <paolo.valente@linaro.org>

# Conflicts:
#	block/bfq-iosched.c

(cherry picked from commit efb2168f3ad6994767feb01346f25b33814b42ed)
(cherry picked from commit 6a55edff0d9cfa656546e69eff7ed10820f0d4ba)