EXPERIMENTAL: sched/fair: Reduce balance interval to 0 if we have a misfit task
It is quite a common occurrence that we have a misfit task (a task
running on a CPU which is close to the capacity limit of that CPU)
in the system AND we have recently done a load balance operation.
Normal load-balance rate limiting means we can leave those tasks on
the restrictive CPU for a few hundred ms before we move them.

This patch reduces the balance interval to zero when there are such
tasks running.

Change-Id: I71d74734d3bd6bca80476c0f466ba48914b7ac02
Signed-off-by: Leo Yan <leo.yan@linaro.org>
Signed-off-by: Chris Redpath <chris.redpath@arm.com>
(cherry picked from commit 41739e5db12dca9fdefb0afea62ea270428b83d3)